+++ src/implementation.jl
@@ -2,7 +2,7 @@ module Implementation
 
 import OhMyThreads: treduce, tmapreduce, treducemap, tforeach, tmap, tmap!, tcollect
 
-using OhMyThreads: chunks, @spawn, @spawnat
+using OhMyThreads: chunks, @spawn, @spawnat, WithTaskLocals, promise_task_local
 using OhMyThreads.Tools: nthtid
 using OhMyThreads: Scheduler, DynamicScheduler, StaticScheduler, GreedyScheduler
 using OhMyThreads.Schedulers: chunking_enabled, chunking_mode, ChunkingMode, NoChunking,
@@ -62,15 +62,17 @@ function _tmapreduce(f,
     if chunking_enabled(scheduler)
         tasks = map(_chunks(scheduler, first(Arrs))) do inds
             args = map(A -> view(A, inds), Arrs)
-            @spawn threadpool mapreduce(f, op, args...; $mapreduce_kwargs...)
+            # Note, calling `promise_task_local` here is only safe because we're assuming that
+            # Base.mapreduce isn't going to magically try to do multithreading on us... 
+            @spawn threadpool mapreduce(promise_task_local(f), promise_task_local(op), args...; $mapreduce_kwargs...)
         end
-        mapreduce(fetch, op, tasks)
+        mapreduce(fetch, promise_task_local(op), tasks)
     else
         tasks = map(eachindex(first(Arrs))) do i
             args = map(A -> @inbounds(A[i]), Arrs)
-            @spawn threadpool f(args...)
+            @spawn threadpool promise_task_local(f)(args...)
         end
-        mapreduce(fetch, op, tasks; mapreduce_kwargs...)
+        mapreduce(fetch, promise_task_local(op), tasks; mapreduce_kwargs...)
     end
 end
 
@@ -84,9 +86,9 @@ function _tmapreduce(f,
     (; nchunks, split, threadpool) = scheduler
     chunking_enabled(scheduler) && auto_disable_chunking_warning()
     tasks = map(only(Arrs)) do idcs
-        @spawn threadpool f(idcs)
+        @spawn threadpool promise_task_local(f)(idcs)
     end
-    mapreduce(fetch, op, tasks; mapreduce_kwargs...)
+    mapreduce(fetch, promise_task_local(op), tasks; mapreduce_kwargs...)
 end
 
 # StaticScheduler: AbstractArray/Generic
@@ -102,16 +104,22 @@ function _tmapreduce(f,
         tasks = map(enumerate(_chunks(scheduler, first(Arrs)))) do (c, inds)
             tid = @inbounds nthtid(mod1(c, nt))
             args = map(A -> view(A, inds), Arrs)
-            @spawnat tid mapreduce(f, op, args...; mapreduce_kwargs...)
+            # Note, calling `promise_task_local` here is only safe because we're assuming that
+            # Base.mapreduce isn't going to magically try to do multithreading on us... 
+            @spawnat tid mapreduce(promise_task_local(f), promise_task_local(op), args...; mapreduce_kwargs...)
         end
-        mapreduce(fetch, op, tasks)
+        # Note, calling `promise_task_local` here is only safe because we're assuming that
+        # Base.mapreduce isn't going to magically try to do multithreading on us... 
+        mapreduce(fetch, promise_task_local(op), tasks)
     else
         tasks = map(enumerate(eachindex(first(Arrs)))) do (c, i)
             tid = @inbounds nthtid(mod1(c, nt))
             args = map(A -> @inbounds(A[i]), Arrs)
-            @spawnat tid f(args...)
+            @spawnat tid promise_task_local(f)(args...)
         end
-        mapreduce(fetch, op, tasks; mapreduce_kwargs...)
+        # Note, calling `promise_task_local` here is only safe because we're assuming that
+        # Base.mapreduce isn't going to magically try to do multithreading on us... 
+        mapreduce(fetch, promise_task_local(op), tasks; mapreduce_kwargs...)
     end
 end
 
@@ -128,9 +136,13 @@ function _tmapreduce(f,
     nt = nthreads()
     tasks = map(enumerate(chnks)) do (c, idcs)
         tid = @inbounds nthtid(mod1(c, nt))
-        @spawnat tid f(idcs)
+        # Note, calling `promise_task_local` here is only safe because we're assuming that
+        # Base.mapreduce isn't going to magically try to do multithreading on us... 
+        @spawnat tid promise_task_local(f)(idcs)
     end
-    mapreduce(fetch, op, tasks; mapreduce_kwargs...)
+    # Note, calling `promise_task_local` here is only safe because we're assuming that
+    # Base.mapreduce isn't going to magically try to do multithreading on us... 
+    mapreduce(fetch, promise_task_local(op), tasks; mapreduce_kwargs...)
 end
 
 # GreedyScheduler
@@ -155,11 +167,15 @@ function _tmapreduce(f,
         end
     end
     tasks = map(1:ntasks) do _
-        @spawn mapreduce(op, ch; mapreduce_kwargs...) do args
-            f(args...)
+        # Note, calling `promise_task_local` here is only safe because we're assuming that
+        # Base.mapreduce isn't going to magically try to do multithreading on us... 
+        @spawn mapreduce(promise_task_local(op), ch; mapreduce_kwargs...) do args
+            promise_task_local(f)(args...)
         end
     end
-    mapreduce(fetch, op, tasks; mapreduce_kwargs...)
+    # Note, calling `promise_task_local` here is only safe because we're assuming that
+    # Base.mapreduce isn't going to magically try to do multithreading on us... 
+    mapreduce(fetch, promise_task_local(op), tasks; mapreduce_kwargs...)
 end
 
 function check_all_have_same_indices(Arrs)
@@ -183,6 +199,25 @@ function tforeach(f, A...; kwargs...)::Nothing
 end
 
 #-------------------------------------------------------------
+
+function maybe_rewrap(g::G, f::F) where {G, F}
+    g(f)
+end
+
+"""
+   maybe_rewrap(g, f)
+
+takes a closure `g(f)` and if `f` is a `WithTaskLocals`, we're going
+to unwrap `f` and delegate its `TaskLocalValues` to `g`.
+
+This should always be equivalent to just calling `g(f)`.
+"""
+function maybe_rewrap(g::G, f::WithTaskLocals{F}) where {G, F}
+    (;inner_func, tasklocals) = f
+    WithTaskLocals(vals -> g(inner_func(vals)), tasklocals)
+end
+
+#------------------------------------------------------------
 
 function tmap(f, ::Type{T}, A::AbstractArray, _Arrs::AbstractArray...; kwargs...) where {T}
     Arrs = (A, _Arrs...)
@@ -229,7 +264,7 @@ function _tmap(scheduler::DynamicScheduler{NoChunking},
     tasks = map(eachindex(A)) do i
         @spawn threadpool begin
             args = map(A -> A[i], Arrs)
-            f(args...)
+            promise_task_local(f)(args...)
         end
     end
     v = map(fetch, tasks)
@@ -244,7 +279,7 @@ function _tmap(scheduler::DynamicScheduler{NoChunking},
         kwargs...)
     (; threadpool) = scheduler
     tasks = map(A) do idcs
-        @spawn threadpool f(idcs)
+        @spawn threadpool promise_task_local(f)(idcs)
     end
     map(fetch, tasks)
 end
@@ -258,7 +293,7 @@ function _tmap(scheduler::StaticScheduler{NoChunking},
     nt = nthreads()
     tasks = map(enumerate(A)) do (c, idcs)
         tid = @inbounds nthtid(mod1(c, nt))
-        @spawnat tid f(idcs)
+        @spawnat tid promise_task_local(f)(idcs)
     end
     map(fetch, tasks)
 end
@@ -272,10 +307,13 @@ function _tmap(scheduler::Scheduler,
     Arrs = (A, _Arrs...)
     idcs = collect(_chunks(scheduler, A))
     reduction_f = append!!
-    v = tmapreduce(reduction_f, idcs; scheduler, kwargs...) do inds
-        args = map(A -> @view(A[inds]), Arrs)
-        map(f, args...)
+    mapping_f = maybe_rewrap(f) do f
+        (inds) -> begin
+            args = map(A -> @view(A[inds]), Arrs)
+            map(f, args...)
+        end
     end
+    v = tmapreduce(mapping_f, reduction_f, idcs; scheduler, kwargs...)
     reshape(v, size(A)...)
 end
 
@@ -290,11 +328,14 @@ end
     end
     Arrs = (A, _Arrs...)
     @boundscheck check_all_have_same_indices((out, Arrs...))
-    tforeach(eachindex(out); scheduler, kwargs...) do i
-        args = map(A -> @inbounds(A[i]), Arrs)
-        res = f(args...)
-        out[i] = res
+    mapping_f = maybe_rewrap(f) do f
+        function mapping_function(i)
+            args = map(A -> @inbounds(A[i]), Arrs)
+            res = f(args...)
+            out[i] = res
+        end
     end
+    tforeach(mapping_f, eachindex(out); scheduler, kwargs...)
     out
 end